+++ src/components/FileName.vue
@@ -3,20 +3,12 @@
       :class="[isActive ? 'text-bold' : '', fileStatus]"
       :data-cy="`file-tab-label-${label}`"
     >
-      {{label}}
+      {{ label }}
     </span>
 </template>
 
 <script setup>
-import FileEvent from 'src/composables/events/FileEvent';
-import {
-  ref,
-  onMounted,
-  onUnmounted,
-  watch,
-} from 'vue';
-import { useRoute } from 'vue-router';
-import { getStatus } from 'src/composables/Project';
+import { ref, watch } from 'vue';
 
 const props = defineProps({
   path: {
@@ -37,36 +29,10 @@ const props = defineProps({
   },
 });
 
-const route = useRoute();
 const fileStatus = ref(props.status);
-let updateFileSubscription;
-
-/**
- * Update fileStatus when file content is updated.
- * @param {String} filePath - Path (id) of the updated file.
- */
-async function onUpdateFile(filePath) {
-  if (filePath === props.path) {
-    const filePathStatus = await getStatus(
-      route.params.projectName,
-      [filePath],
-      (f) => f === filePath,
-    );
-
-    fileStatus.value = filePathStatus[0].status;
-  }
-}
 
 watch(() => props.status, () => {
   fileStatus.value = props.status;
-});
-
-onMounted(() => {
-  updateFileSubscription = FileEvent.UpdateFileEvent.subscribe(onUpdateFile);
-});
-
-onUnmounted(() => {
-  updateFileSubscription.unsubscribe();
 });
 
 </script>