+++ src/components/tab/FileTabHeader.vue
@@ -1,0 +1,58 @@
+<template>
+  <q-tab
+    :name="file.id"
+    no-caps
+    :ripple="false"
+    :class="['tab-item', { 'active': isActive }]"
+    :data-cy="`${isActive? 'active' : 'inactive'}-tab`"
+  >
+    <div class="row items-center">
+      <file-name
+        :path="file.id"
+        :isActive="isActive"
+        :label="file.label"
+        :status="file.information?.status"
+      />
+      <q-btn
+        flat
+        round
+        size="xs"
+        :color="isActive? 'primary' : 'gray'"
+        icon="fa-solid fa-xmark"
+        class="q-ml-sm"
+        @click.stop="emit('update:close-file', file.id)"
+        data-cy="close-file-tab"
+      />
+    </div>
+  </q-tab>
+</template>
+
+<script setup>
+import FileName from 'src/components/FileName.vue';
+
+const emit = defineEmits(['update:close-file']);
+
+defineProps({
+  file: {
+    type: Object,
+    required: true,
+  },
+  isActive: {
+    type: Boolean,
+    required: true,
+  },
+});
+
+</script>
+
+<style lang="scss" scoped>
+  .tab-item {
+    border-right: solid 1px #eee;
+    background: #eee;
+
+    &.active {
+      background: white;
+      box-shadow: inset 0 -1px 0 white;
+    }
+  }
+</style>